% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount_exp.R
\name{discount_exp}
\alias{discount_exp}
\title{The Exponential Discounting Model}
\usage{
discount_exp(A, b, X)
}
\arguments{
\item{A}{Amount of reward}

\item{b}{Discounting rate parameter}

\item{X}{Independent variable (i.e., delay, cognitive effort, probability)}
}
\value{
Subjective value of the reward of amount A
}
\description{
The model was obtained from the conference presentation:
Leonard Green, "The Discounting Function" SQAB
The presentation can be viewed on Youtube:
https://www.youtube.com/watch?v=CDto8pAyG5M
}
\details{
The notation was obtained from:
Myerson, J., Green, L., & Warusawitharana, M. (2001). Area under the curve as a measure of discounting. Journal of the experimental analysis of behavior, 76 2, 235-43 .
}
\examples{
A <- 5000
b <- 7.760
X <- .5
discount_exp(A = A, b = b, X = X)
}
