% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount_hypl.R
\name{discount_hypl}
\alias{discount_hypl}
\title{The Hyperbolic-Like Discounting Model}
\usage{
discount_hypl(A, b, X, s)
}
\arguments{
\item{A}{Amount of reward}

\item{b}{Discounting rate parameter}

\item{X}{Independent variable (i.e., delay, cognitive effort, probability)}

\item{s}{Non-linear scaling factor (set s = 1 to run the hyperbolic discounting model)}
}
\value{
Subjective value of the reward of amount A
}
\description{
The Hyperbolic-Like Discounting Model
}
\examples{
A <- 5000
b <- 7.760
X <- .5
s <- 2
discount_hypl(A = A, b = b, X = X, s = s)
}
