% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traper.R
\name{traper}
\alias{traper}
\title{Traper}
\usage{
traper(df, x, y)
}
\arguments{
\item{df}{A data frame with one x and one y value on each line}

\item{x}{Column of data frame containing trial index (i.e., successive delays,
cognitive effort, probability).  traper function sorts the data by x.}

\item{y}{Column of data frame containing subjective values (or scores)}
}
\value{
Data frame where each row contains the data to compute the area of
one trapezoid. The *_lead columns are the x and y values from the next row.
}
\description{
Convert df with one x and one y value per row into trapezoidal data where x
values represent successive indexes (i.e., successive delays, cognitive effort,
probability) and y values represent subjective values associated with the delay.
}
\details{
Citation: Myerson, J., Green, L., & Warusawitharana, M. (2001). Area under the curve as a measure of discounting. Journal of the experimental analysis of behavior, 76 2, 235-43 .
}
\examples{

#Create data frame with successive delays and subjective values
df_test <- data.frame(delay = sample(1:100, 50, replace = FALSE),
                      value = sample(1:100, 50, replace = TRUE))

df_test_2 <- traper(df_test, x = delay, y = value)
}
